requirements.txt
fastapi
uvicorn
jinja2
httpx
python-multipart
python-dotenv
